Change Contest from ManyToMany -> ForeignKey in Problems
We need the problem to exist in one contest at a time,
moving will still make it part of another competition,
but a single. Hence the decision. Besides it makes it easy
to backtrack from Problem to a URL, which will otherwise be
not-possible because of multiple existence of the problem.