Bulk migrate interface; change home page design
This is a big design decision. We have migrated to a much cooler
design which gives more space for utilization. Besides the color
combinations are much professional. Also host of CSS are added which
should collapse for easier management later!

Also, the home page is now changed from the simple Points view to
a richer Updates section. Most of it is open to change and alteration.