Modify tutorial behavior on problems list page
Earlier the problem list had tutorial if it existed. Firstly it might
happen that an admin could put up a problem and subsequently want to
write a tutorial. While he is in midst, it might not be necessary to
publish it. Hence, currently a "is_visible" option is added to tutorial
model to tell if the tutorial is visible to public. Also if it is not
checked, an admin will still be able to view all tutorials.